+++ MovieRecommendationBot/src/main/java/ru/spbstu/movierecbot/dao/WatchListDao.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface WatchListDao {
     //Добавление фильма в список "Буду смотреть"
-    int addToWatchlist(int telegramId, int filmId, String title);
+    int addToWatchlist(long telegramId, int filmId, String title);
     //Получение списка всех фильмов "Буду смотреть"
-    List<WatchListRecord> getWatchlistByTelegramId(int telegramId);
+    List<WatchListRecord> getWatchlistByTelegramId(long telegramId);
     //Удаление фильма по id из списка "Буду смотреть"
-    int deleteFromWatchList(int telegramId, int filmId);
+    int deleteFromWatchList(long telegramId, int filmId);
 }
